Fall back to system default PKG_CONFIG_PATH if unset
I'm a user of c-for-go. Unfortunately it couldn't work by default unless
a sensible PKG_CONFIG_PATH is specified, and it isn't totally obvious
how to determine that.

Instead, if PKG_CONFIG_PATH is unset, try to invoke the system pkg-config
and ask what path to use. If it's unavailable, then we just fall back to
the old behaviour of just returning an error.

For reference, here's what pkg-config returns on my system:

$ pkg-config --variable=pc_path pkg-config
/usr/local/lib/x86_64-linux-gnu/pkgconfig:/usr/local/lib/pkgconfig:/usr/local/share/pkgconfig:/usr/lib/x86_64-linux-gnu/pkgconfig:/usr/lib/pkgconfig:/usr/share/pkgconfig